code/requirements.txt
epigenomic_dataset~=1.1.5
pandas~=1.0.4
ucsc_genomes_downloader~=1.1.18
numpy~=1.18.4
keras_bed_sequence~=1.1.1
scikit-learn~=0.23.1
tensorflow~=2.2.0
matplotlib~=3.2.1
Boruta~=0.3
scipy~=1.4.1
tqdm~=4.46.0
prince~=0.7.0
sanitize_ml_labels~=1.0.9
barplots~=1.1.4
Pillow~=7.1.2